requirements.txt
biopython>=1.79
pandas>=1.3.0
numpy>=1.20.0
rich>=12.0.0
requests>=2.25.0
